components/ico/buy, view: modularize
